Add REST endpoint to get the code owner statuses for the files in a change
This REST endpoint returns the code owner statuses for the the files in
a change. Depnding on the change type changed files can have a new
and/or old path (e.g. a rename has a new and old path, a deletion only
has an old path) for which code owner approvals are required. To make
callers aware of this the returned file code owner statuses contain code
owner statuses for the old and new path, if they are present.

A path in a change can have the following code owner statuses:

* INSUFFICIENT_REVIEWERS:
  The path needs a code owner approval, but none of its code owners is
  currently a reviewer of the change.

* PENDING:
  A code owner of this path has been added as reviewer, but no code
  owner approval for this path has been given yet.

* APPROVED:
  The path has been approved by a code owner or a code owners override
  is present.

The GetCodeOwnerStatuses REST endpoint is offered on change level,
rather than on revision level because:

* The code owner statuses are based on votes and we always show the
  current votes on the change screen even if users view an old revision
  / patch set (it would be confusing for users to see the current votes,
  but non-matching code owner statuses).

* The code owner statuses for old revisions are never relevant and may
  confuse users (e.g. the code owner status 'PENDING' means that we are
  waiting for the code owner approval of a current reviewer, but the UI
  doesn’t allow reviewers to vote on non-current revisions, so that this
  approval can never happen).

This means the frontend can only show owner statuses for the current
patch set.

In addition there is a risk for races if new patch sets are uploaded:

1. The frontend loads the change when patch set N is current.
2. A user uploads a new patch set N+1.
3. The frontend loads the owner statuses which are returned for patch
   set N+1.

To allow the frontend to detect this situation, the response includes
the patch set for which the owner statuses were returned (see field
patch_set_number). The frontend can then show a warning or do a refresh
when the patch sets mismatch.

To avoid this issue we considered to return the owner statuses as part
of ChangeInfo (plugins field), but we know from the checks plugin that
this is not working well:

* the code owner statuses should not be returned by default, but only
  when requested (most callers don’t need the code owner statuses)

* this means the caller would need to set an option on the request to
  include the code owner statuses

* the frontend loads the ChangeInfo before it knows which plugins are
  active, hence at the moment when the ChangeInfo is loaded the frontend
  doesn’t know yet that it needs the code owner statuses

We also considered returning the file code owner statuses as map of file
path to file code owner status, but this doesn't work well when a file
is renamed, as in this case the files has 2 code owner statuses, one for
the old path and one for the new path. We could have returned both path
in the map but then the client would have needed to figure out which
2 paths form a file rename.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6c449448e9d541a6d36ea9ac4805248a6cd6ed60