Assume implicit approval from patch set uploader for owned files
If the uploader of the current patch set owns any of the touched files,
these files are now automatically considered as approved by the patch
set uploader.

This uses the uploader of the current patch set, rather than the
committer or author because the uploader cannot easily be forged.
Committer and author can be forged easily if the user has the
corresponding permissions and also forging them is normal for some use
cases (e.g. the author is forged if a user uploads a patch set to a
change of another user, for example while the change author is in
vacation). If we would assume implicit approvals from the committer or
author this would be unsafe, since this way anyone who can forge
committer or author could approve file modifications in their name by
forging their identity. This is also the reason why the
ignoreSelfApproval rule on label types [1] prevents self approval from
the uploader rather than from the committer or author.

[1] https://gerrit-review.googlesource.com/Documentation/config-labels.html#label_ignoreSelfApproval

Change-Id: I58efd122748f14b385d21a447e93fc27c076d5c6
Signed-off-by: Edwin Kempin <ekempin@google.com>