Do not fail with ISE if plugin configuration is invalid
The code-owners plugin has several configuration parameters that control
the behaviour. If any of the required parameters is set to an invalid
value the code owners functionality cannot be provided and we must fail.

So far we threw an IllegalStateException if an invalid configuration
value was found. This is not good since the IllegalStateException's
result in internal server errors and thus a mistake in configuration can
cause many internal server errors which will cause an SLO violation.

Instead we now throw a new InvalidPluginConfigurationException which is
a checked exception that needs to be handled by the callers. The REST
endpoints catch and convert these exceptions into a '409 Conflict'
responses. 4xx responses are usually used for client errors which is
strictly speaking not the case here. However Gerrit core uses '409
Conflict' in similar cases (e.g. if an invalid submit rules was
configured) and we want to be consistent with this.

Since we have now a second class that is related to the plugin
configuration, we start a new package for config-related classes.

While we are at it, we also slightly improve the error messages in case
of invalid configuration (we now include the project name if an invalid
configuration parameter was read from a project-level config file).

Lastly, we now have explicit statements to log the
InvalidPluginConfigurationExceptions. This is necessary since this
exception is no longer resulting into an internal server error which was
logged automatically.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia1ef91e40ebe4dc2054163c6f53ef214b039fab0