Move code to read required approval config into RequiredApproval class
This allows use to have unit tests for these methods and reduces the
size of the CodeOwnersPluginConfiguration class (if we continue adding
methods for each new config parameter to CodeOwnersPluginConfiguration
this class will quickly grow too much). Also we will add code to
validate the required approval configuration to RequireApproval in a
later change and the reading and validation code should be next to each
other.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia75341b9861d8c6cde1b0697a43c68b38a0eb4be