Move code to read backend configuration into a separate class
This allows use to have unit tests for these methods and reduces the
size of the CodeOwnersPluginConfiguration class (if we continue adding
methods for each new config parameter to CodeOwnersPluginConfiguration
this class will quickly grow too much). Also we will add code to
validate the backend configuration to new Backend class in a later
change and the reading and validation code should be next to each other.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia3e7757b468f1418f1a97a3a9dfffc7e27b57faa