Validate modifications to code-owners.config file on push
Invalid values in the code-owners.config file in refs/meta/config can
break the the whole code owners functionality for this project and its
child projects. To prevent this from happening, validate the
code-owners.config file on push when it is modified so that the push is
rejected if the code-owners.config is invalid or contains invalid
values.

Change-Id: Ib037028a52c29000c534344ba8aa9870890b1a22
Signed-off-by: Edwin Kempin <ekempin@google.com>