Make InvalidPluginConfigurationException a runtime exception
Having InvalidPluginConfigurationException as a checked exception has
several disadvantages:

- many methods need to declare it as being thrown
- we can't call methods that throw InvalidPluginConfigurationException
  in lambda expressions

The exception was made a checked exception so that the REST endpoints
are required to handle it properly and map it to a conflict response.
However this is also not nice, since it means that we need to do the
same exception handling in all REST endpoints.

Instead we make InvalidPluginConfigurationException a runtime exception
now and we register an ExceptionHook to convert
InvalidPluginConfigurationException to conflict responses. This way the
conversion is done in a single place and does not need to be done in
each REST endpoint.

Change-Id: Ib89f99b4c65d5988ec6554774e80978bde393e8d
Signed-off-by: Edwin Kempin <ekempin@google.com>