Add submit rules that checks code owner approals
Changes with insufficent code owner approvals should not be submittable.
To achieve this we add a new submit rule that checks the code owner
approvals on submit and prevents the submission of changes with
insufficent code owner approvals.

Whether a change has sufficient code owner approvals is computed based
on the code owner statuses for the files that are touched in the change.

To know whether a change has insufficent code owner approvals we do not
need to compute the code owner statuses for all touched files, but we
can abort as soon as we find a file that has a non-approved status.

In order to abort early the logic in CodeOwnerApprovalCheck is changed
to produce a stream of FileCodeOwnerStatus's rather than a set of
FileCodeOwnerStatus's.

Change-Id: Id8db8852502c877fd384972a6def7399eec154b4
Signed-off-by: Edwin Kempin <ekempin@google.com>