Allow projects and branches to opt-out of the code owners functionality
For projects and branches it should be possible to opt-out of the code
owners functionality. If the code owners functionality is disabled for a
project or branch submitting changes for them doesn't require code owner
approvals.

The code owners REST API fully works for disabled projects and branches.
This is useful to test the code owners functionality while not enforcing
code owner approvals yet.

For branches it should be possible to specify exact refs, ref patterns
and regular expressions to define that matching branches opt-out.

Whether the code owners functionality is disabled for a project/branch
is returned to the frontend as part of the response of the
GetCodeOwnersProjectConfig REST endpoint so that the frontend knows when
to disable/hide code owner UI controls.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I24324a84008fc785175e463395c27663a2608d72