Allow projects and branches to opt-out of the code owners functionality
For projects and branches it should be possible to opt-out of the code
owners functionality. If the code owners functionality is disabled for a
project or branch the code owners API is disabled and submitting changes
doesn't require code owner approvals.

For branches it should be possible to specify exact refs, ref patterns
and regular expressions to define that matching branches opt-out.

The following REST endpoints are disabled and return '405 Method Not
Allowed' if the code owners functionality is disabled for the request
resource:

* Get code owners for path in branch/change
* Get code owner statuses

Getting code owner configs is still working.

Whether the code owners functionality is disabled for a project/branch
is returned to the frontend as part of the response of the
GetCodeOwnersProjectConfig REST endpoint so that the frontend knows when
to disable/hide code owner UI controls.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I24324a84008fc785175e463395c27663a2608d72