Mark GetCodeOwnerConfig REST endpoint as experimental
The GetCodeOwnerConfig REST endpoint returns a code owner config as
JSON. The JSON structure matches the proto structure that is used by the
Proto backend. The proto is not fully defined yet and we expect that it
will still change. Any change should be refelected back to the
CodeOwnerConfigInfo. This means CodeOwnerConfigInfo is not stable yet
and no productive callers should rely on it. Hence mark the
GetCodeOwnerConfig REST endpoint that returns CodeOwnerConfigInfo as
experimental to make callers aware of this. There is currently no
stakeholder that needs this REST endpoint, hence it's okay that this
REST endpoint is only experimental for now. A follow-up change will add
a config option that needs to set to enable experimental REST endpoints,
so that this REST endpoint will be disabled by default so that no
productive callers use it by accident.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7e14903cc6ccd26e0a47e306f47859e3833a4a8f