RequiredApproval: Extract config code into separate class
The RequiredApproval class currently has 2 purposes it knows how to read
and validate the required approval config, and it also represents the
loaded/parsed required approval. Extract the config code into a new
RequiredApprovalConfig class. This will allow us to reuse the config
code for other config parameters that also represent a required approval
(e.g. it should be possible to configure which approval is required to
override code owners). In addition the RequiredApprovalConfig class will
now be able to use injection (this was not possible with
RequiredApproval which is an AutoValue class). Using injection in
RequiredApprovalConfig will be done in a follow-up change to make
reviewing easier.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib4d62fdce922b2140ebcddf208c2f4f4338cf010