Add config for the approval to override the code owners submit check
It should be possible to override the code owners submit check by voting
on a special label. Which label and voting value is required for an
override should be configurable. This change adds a configuration option
for this, but doesn't implement support for overrides yet (this will be
implemented by a follow-up change).

To avoid code duplication we reuse the parsing and validation logic for
the required approval config to parse and validate the override approval
config.

A difference is that for the override approval there is no hard-coded
default, but if no override configuration is available overrides are
simply disabled. Since the override functionality is optional we can
also gracefully handle configuration errors. If there is an invalid
configuration for the override approval, we log a warning and the
override functionality is disabled.

Which override approval is configured is exposed to the frontend via the
GetCodeOwnerProjectConfig REST endpoint. This allows the frontend to
detect overrides and have specific behaviour if they are present.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If2c155021b842d2e45a10820c6330bc0954c03bd