Support overriding of the code owners submit check by special approval
If an approval on the configured override label is present, all files
should show up as approved.

We decided against returning an explicit "OVERRIDDEN" status in this
case to keep the number of possible statuses low and to not couple the
GetCodeOwnerStatus REST endpoint with the override feature.

Since callers know which approvals are present on a change, and they can
look up which approval counts as override approval, they can find out
that a changes has an override if they want.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib2df8e70e4127821fc0a0e4283da9424d95f9700