Remove deprecation warning from FindOwnersBackend
Initially the FindOwnersBackend was only intended to be used during the
migration period to the new proto-based OWNERS_METADATA format that is
supported by the ProtoBackend. It turned out now that defining a stable
version of the proto format will take much longer, hence we changed our
plans and decided to stick with the current OWNERS format for now. This
means a migration to the proto-based format will not happen any time
soon.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I36253efdcd89b012d2a35c890d37b8dd860d9f93