AbstractCodeOwnerConfigParserTest: Pass full CodeOwnerConfig to subclasses
The subclasses of AbstractCodeOwnerConfigParserTest must provide the
expected format string for a CodeOwnerConfig. Instead of providing the
CodeOwnerConfig class to the subclasses, we provided them the data that
is contained in a CodeOwnerConfig (includeParentCodeOwners flag + code
owner sets). This way would have needed to add another parameter each
time CodeOwnerConfig is extended. Simplify this by providiing the
CodeOwnerConfig to the subclasses.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I641ad18b24aa41dee388180d554cb9bd79d28456