Rename LocalCodeOwners to PathCodeOwners
The LocalCodeOwners class computes the code owners for a path that are
locally defined in a given code owner config. Locally here meant that
only code owners that are mentioned in that code owner config were taken
into account, while inherited and imported code owners were not
considered. We are now going to support importing code owner configs and
we want to compute the imported code owners in this class too. Hence the
name LocalCodeOwners will not be suitable anymore. This is why we rename
it to PathCodeOwners, a class that computes the code owners from a code
owner config for a given path.

The term local code owners also started to become confusing after we
named code owners that apply to the whole folder global code owners
(code owners from code owner sets without path expressions), as it's
hard to understand how global code owners can also be local code owners
(these are global code owners that are locally defined in the code owner
config).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ic24631bac67983c1ea18ce40dfd4bba6a7b308e1