Change PathCodeOwners to wrap code owner config and path
Instead of having PathCodeOwners as singleton and needing to pass in the
code owner config and path to all methods, change PathCodeOwners to be
non-singleton and have the code owner config and the path as members.

This simplifies the code and allows us to add further factory methods
later, e.g. to instantiate PathCodeOwners for a reference to an imported
code owner config.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I75ccc98e30901fb806b62d1f4b4160e8d14a2d40