Allow to create PathCodeOwners from code owner config key
This way callers that only have the code owner config key do not need to
load the code owner config themselves which makes the call a little
easier for them. For PathCodeOwners we will anyway need a dependency to
CodeOwners later to be able to load imported code owner configs.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5699ce8df38523dd8f4e508c9a1d8568082fc530