Add global imports to code owner configs
This change adds a field for global imports to the CodeOwnerConfig
class. Populating and reading this field will be implemented in
follow-up changes, as well as support for per-file imports (imports that
apply only for given path expressions).

Code owner configs can import other code owner configs from:
1. a different path in the same project
2. any path in another project (same host)

Being able to specify the branch from which the code owner config is
imported is desired too, but will be implemented only later (we first
want to cover the import functionality that is currently supported by
the find-owners plugin and the find-owners plugin doesn't support
specifying branches for imports).

Paths of imported code owner configs can be relative or absolute.

The find-owners plugin currently supports 2 import modes:
1. using the "file" directive (e.g. "file: <file>"):
   only the global code owners from the imported code owner config are
   considered (code owners from code owner sets without path
   expressions)
2. using the "include" directive (e.g. "include <file>"):
   imports global and per-file code owners (code owners from all code
   owner sets) and also the "set noparent" flag

Details about these import modes are described at [2].

Since we need to be compatible with the find-owners plugin we need to
support these import modes too. In the code we represent them via the
CodeOwnerConfigImportMode enum. The CodeOwnerConfigImportMode exposes a
bunch of methods to control the import behaviour. This way we can easily
tweak the behaviour later.

Imports are represented as CodeOwnerConfigReference's and are stored in
CodeOwnerConfig, instead of loading and resolving them immediately on
parse of the code owner config. This way we can resolve the imports in
the generic layer rather than in the backend layer. This is important
since imports from other projects are possible and these other projects
may use different code owner backends, and the logic to lookup the code
owner backend for a project lives in the generic layer.

[1] https://gerrit.googlesource.com/plugins/find-owners/+/73bb362fd84b2089e2625761253852cc7716a14b/src/main/resources/Documentation/syntax.md

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I46005fbab6e127de66ac38d3d6de076e36cf3f3e