FindOwnersBackend: Support parsing global imports
The find-owners syntax has 2 directives to import code owner configs:
1. the "file" directive (e.g. "file: <file>"):
   only the global code owners from the imported code owner config are
   considered (code owners from code owner sets without path
   expressions)
2. the "include" directive (e.g. "include <file>"):
   imports global and per-file code owners (code owners from all code
   owner sets) and also the "set noparent" flag

Note that there is no ':' after the 'include' keyword.

For both directives it is possible to optionally specify a project in
front of the file name: "file: <project>:<file>" and "include
<project>:<file>"

Note that the documentation at [1] states that the "file" directive
doesn't support projects, but the code in the find-owners plugin
actually does (even if not, we would still want to support this).

[1] https://gerrit.googlesource.com/plugins/find-owners/+/73bb362fd84b2089e2625761253852cc7716a14b/src/main/resources/Documentation/syntax.md

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I183db0b0dede13e999d248ac422e5010db7a2dc1