Allow to optionally specifiy a file name in CodeOwnerConfig.Key
So far CodeOwnerConfig.Key only contained the path of the code owner
config and the file name was determined by the code owner backend (e.g.
the FindOwnersBackend always uses 'OWNERS' as file name). We now allow
to specify a custom file name and the code owner backends can decide
whether they support this file name. This will be useful for resolving
imports of code owner configs, where the reference to the imported code
owner config contains the name of the file that should be imported. On
import we should import exactly this file, or ignore the import if the
file name is unsupported.

For now, all code owner backends only support their default file names.
However in future we want to support file names with arbitrary
extensions ("<default-name>_<extension>", e.g. "OWNERS_BUILD").

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I4af068256d829f2b4f5b5c5936175b687f29c1f5