CodeOwnerConfig: Add field for the revision from which the config was read
This makes debugging easier in case something is wrong with a code owner
config, as it allows us to inspect the raw code owner config file in the
commit.

Being able to tell from which revision an object was loaded is a best
practice in Gerrit core that we want to follow here too.

Except in test code, CodeOwnerConfigs are always loaded from a file,
hence there is always a revision from which the code owner config was
loaded, which is why we make the revision a mandatory field. In unit
tests we sometimes just instantiate CodeOwnerConfigs which requires a
little more overhead now, since we need to set a fake revision, but this
doesn't seem to be too bad.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I3b2463cf159bff00bbdb354f9effddcdd29f883a