Allow to import code owner config files with post-/pre-fix in their name
So far the names of code owners config files must always match the
default/configured file name and imports of code owner config files with
other file names were ignored. With this change we now relax this a bit
and allow to import code owner configs that use a post- or pre-fix in
their name, e.g. 'OWNERS_<post-fix>' or '<pre-fix>_OWNERS'.

Code owner config files with post- or pre-fix in their name only have an
effect if they are imported into a regular code owner config file and
are otherwise ignored.

Code owner config files with post- or pre-fix can be recognized as code
owner config files on push so that we will be able to validate them when
they are modified (validation on push will only be implemented by
follow-up changes). This would not be possible if we would allow to
import code owner config files with arbitrary names, which is why we
still do not allow arbitrary names.

Having code owner config files with post- or pre-fix is a common
practice for users of the find-owners plugin that we need to support
when we want to migrate users of the find-owners plugin to the
code-owners plugin.

Code owner config files with post- or pre-fix are used to define
commonly used sets of code owners that are defined once and then
imported into multiple code owner configs instead of redefining the code
owners in each of the code owner configs and risking that they get
inconsistent. Basically it's kind of a work-around for not being able to
use groups in code owner configs, but adding support for groups will not
be added any time soon (even if groups were supported our stakeholders
would still want this feature).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id69c34117241b625a4b90a132c88ea96ce80e522