Allow to specify branch for imports
So far code owner configs are always imported from the same branch in
which the importing code owner config is stored. Now we allow to specify
the branch for imports explicitly. If no branch is specified we still
import the code owner config from the same branch in which the importing
code owner config is stored.

In the find-owners syntax imports are specified as:

  file: <file-path>

or

  file: <project-name>:<file-path>

Now in addition we support:

  file: <project-name>:<branch>:<file-path>

This means if a branch is specified it is mandatory to also specify the
project since 'file: <branch>:<file-path>' would not be distinguishable
from 'file: <project-name>:<file-path>', but we don't see this as a
problem.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I310afcc6690c45f371ec905c6599ac8eb70fb1e6