Consider project owners as code owner if there is no code owner config yet
If a project or branch is newly created there are no code owners yet
(since there are no code owner config files yet). Trying to add an
initial code owner config would require code owner approval which cannot
be granted since there are no code owners yet (submitting the initial
code owner config would require an code owner override). To avoid this
situation we consider project owners as code owner if (and only if)
there is no code owner config yet. This way changes done by code owners
do not require an explicit code owner approval (since a code owner
approval from the uploader of the current patch set is always implicit)
and changes from other users can be approved by a project owner.

With this change we do not yet suggest project owners as code owners
when we are in bootstrapping mode. Implementing this is not trivial
(since project ownership is assigned to groups and we cannot list
members of external groups) which is why we are not implementing this
for now, but it may be implemented later.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If1dc5d3fde93f8c7e6a7d1d28e55ea94fb34e093