CodeOwnerProjectConfigInfo: Always populate status field
If we want to add further fields to the status it gets impractical to
set the status only if at least one of its fields is set. Always setting
the status, even if it's empty, simplifies the coding and also makes
reading the fields easier for caller. In addition this makes the status
field consistent with the general field which is also always set.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib41f0c61e7ad7fae91ebd52472f4207a9d2cee90