Reject non-parseable code owner config files on push
If there are non-parseable code owner config files in the repository
changes that touch files for which these code owner configs are relevant
cannot be submitted (since the code owners cannot be computed). To
prevent this from happening we now disallow adding non-parseable code
owner configs and reject them directly on push. Messages that describe
the validation errors are included in the push messages so that the user
can find out what needs to be fixed.

With this change we only ensure that the code owner config files are
parseable. In follow-up changes we will add further validation that will
be done on the parsed code owner config files (see todo in
CodeOwnerConfigValidator).

If the code-owners plugin configuration is invalid we cannot get the
code owners backend and hence we are not able to detect and validate
code owner config files. Instead of failing in this case (which would
block all change uploads) we only log a warning and accept that it's
possible to add invalid code owner configs while the plugin
configuration is invalid.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I96148b6a5797551ee60e4f96be45a05bded25c31