Allow to upload non-parseable code owner config if it was broken before
If a code owner config is non-parseable we now always allow to update
it, even if the new version is still non-parseable. It's fine to allow
this since this update doesn't do any more damage than what we had
before the upload.

Updating non-parseable code owner configs without fixing the parsing
issues can be useful in some case, e.g. there may be a bot that is
replacing emails in all code owner configs (and this may still be
possible even if code owner configs are not parseable).

In general we only want to block uploads of code owner configs that
introduce issues newly, e.g. if there are 2 issues and an upload fixes
one of them, the upload should succeed. Whether we only look at newly
introduced issues or all issues of a code owner config is controlled by
the new deltaValidation flag in CodeOwnerConfigValidator. This is a flag
because later we also want to offer on-demand validation of checked-in
code owner config files and in this case we do not want to do delta
validation, but find all issues.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I49ad294b83d5b04816543f50fa044738b412cb99