Allow to upload non-parseable code owner config if it was broken before
If a code owner config is non-parseable we now always allow to update
it, even if the new version is still non-parseable. It's fine to allow
this since this update doesn't do any more damage than what we had
before the upload.

Updating non-parseable code owner configs without fixing the parsing
issues can be useful in some case, e.g. there may be a bot that is
replacing emails in all code owner configs (and this may still be
possible even if code owner configs are not parseable).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I49ad294b83d5b04816543f50fa044738b412cb99