Extract test methods to format code owner configs into a separate class
The parser tests know how to format a code owner config so that they can
verify the output of the production formatters. Extract this logic into
a separate class so that it can be reused by other tests. The new
TestCodeOwnerConfigFormatter class automatically uses the syntax of the
configured code owner backend.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia6d64d08737afc0d96bf80db1855256e9ef831bf