Validate code owner references on upload of code owner configs
Prevent that non-resolveable code owner references are added to code
owner configs. E.g. this is helpful to detect typos in email addresses
early.

If the uploader cannot see the account of a user that is being added as
code owner, the upload is rejected since for this uploader the code
owner cannot be resolved. This is important so that uploaders cannot
probe emails for existence (e.g. they cannot add an email and conclude
from the error message whether the email exists).

If a code owner config is non-parseable and a user updates it so that it
becomes parseable, we allow the update even if the new version of the
code owner config contains issues. This is allowed because a parseable
code owner config with issues is better than a non-parseable code owner
config. This allows users to fix parsibility issues without needing to
fix all other issues as well.

Issues only block the upload if they are newly introduced. This means if
an issue already existed in the previous version of the code owner
config it is not considered as a blocking error, but only as a
non-blocking warning. This allows users to fix a subset of issues in
code owner config files (e.g. if a code owner config has 2 issues, users
are able to fix 1 issue and leave the other issue unfixed).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib35955a1c9cd2eef4b35621edeb0979734207e95