Allow to configure allowed email domains
It can be a security issue if emails, and hence the code ownerships that
are assigned to them, are reassigned to new users. Whether reassigning
an email to a new user is possible depends on the email provider (e.g.
if a user unregisters from their email provider, the email provider may
decide to assign the email to a new user). To prevent this issue, this
configuration option allows to limit the allowed email domains for code
owners to trusted email providers that assign emails to users
permanently and never reuse emails for new users. Internally we will use
this to limit code ownerships to @google.com and @chromium.org email
addresses, since we know that these emails cannot be reassigned to new
users. Emails in owner configuration files that do not match this
setting are rejected on update and ignored for the code owner suggestion
and the code owner submit check if they exist.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I942db2e1535f3c81d20bda9b3ea7f9bfb08aec0d